Complete!
CPU passes all 4 of the standard test ROMS.

Next stop Space Invaders!

Merging with main branch

+++ i8080.h
@@ -32,7 +32,7 @@ public:
 	bool halted;
 	bool error;
 	bool interrupt_enabled;
-	int cycles;
+	unsigned long cycles;
 	uint8_t opcode;
 
 	// Emulator Enumerations
